red-teaming/requirements.txt
anthropic
inspect-ai
python-dotenv
slack-sdk
tenacity
pandas
pyarrow
